docs use _app.js
