+++ .github/ISSUE_TEMPLATE/1.bug.md
@@ -14,17 +14,17 @@ about: Create a bug report for Material-UI
 - [ ] This is not a v0.x issue. <!-- (v0.x is no longer maintained) -->
 - [ ] I have searched the [issues](https://github.com/mui-org/material-ui/issues) of this repository and believe that this is not a duplicate.
 
-## Expected Behavior
+## Expected Behavior 🤔
 <!---
     Describe what should happen.
 -->
 
-## Current Behavior
+## Current Behavior 😯
 <!---
     Describe what happens instead of the expected behavior.
 -->
 
-## Steps to Reproduce
+## Steps to Reproduce 🕹
 <!---
     Provide a link to a live example (you can use codesandbox.io) and an unambiguous set of steps to reproduce this bug.
     Include code to reproduce, if relevant (which it most likely is).
@@ -45,13 +45,13 @@ Link:
 3.
 4.
 
-## Context
+## Context 🔦
 <!---
     What are you trying to accomplish? How has this issue affected you?
     Providing context helps us come up with a solution that is most useful in the real world.
 -->
 
-## Your Environment
+## Your Environment 🌎
 <!---
     Include as many relevant details about the environment with which you experienced the bug.
     If you encounter issues with typescript please include version and tsconfig.
@@ -59,7 +59,7 @@ Link:
 
 | Tech         | Version |
 |--------------|---------|
-| Material-UI  | v1.?.?  |
+| Material-UI  | v3.?.?  |
 | React        |         |
 | Browser      |         |
 | TypeScript   |         |
