+++ .github/ISSUE_TEMPLATE/2.feature.md
@@ -14,23 +14,23 @@ about: Create feature request for Material-UI
 - [ ] This is not a v0.x issue. <!-- (v0.x is no longer maintained) -->
 - [ ] I have searched the [issues](https://github.com/mui-org/material-ui/issues) of this repository and believe that this is not a duplicate.
 
-## Expected Behavior
+## Expected Behavior 🤔
 <!---
     Describe how it should work.
 -->
 
-## Current Behavior
+## Current Behavior 😯
 <!---
     Explain the difference from current behavior.
 -->
 
-## Examples
+## Examples 🌈
 <!---
-    Provide a link to the Material design specification, other implementations, 
+    Provide a link to the Material design specification, other implementations,
     or screenshots of the expected behavior.
 -->
 
-## Context
+## Context 🔦
 <!---
     What are you trying to accomplish? How has the lack of this feature affected you?
     Providing context helps us come up with a solution that is most useful in the real world.
