+++ docs/src/modules/components/bootstrap.js
@@ -1,3 +1,10 @@
 import { install } from '@material-ui/styles';
 
+// Disable auto highlighting
+// https://github.com/PrismJS/prism/issues/765
+if (process.browser) {
+  window.Prism = window.Prism || {};
+  window.Prism.manual = true;
+}
+
 install();
