+++ pages/_document.js
@@ -1,6 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import Document, { Head, Main, NextScript } from 'next/document';
-import getPageContext from 'docs/src/modules/styles/getPageContext';
 
 // You can find a benchmark of the available CSS minifiers under
 // https://github.com/GoalSmashers/css-minification-benchmark
@@ -65,7 +65,7 @@ class MyDocument extends Document {
             dangerouslySetInnerHTML={{
               __html: `
 window.ga=window.ga||function(){(ga.q=ga.q||[]).push(arguments)};ga.l=+new Date;
-window.ga('create', '${GOOGLE_ID}', 'auto');
+window.ga('create','${GOOGLE_ID}','auto');
               `,
             }}
           />
@@ -94,11 +94,20 @@ MyDocument.getInitialProps = async ctx => {
   // 1. page.getInitialProps
   // 3. page.render
 
-  // Get the context of the page to collected side effects.
-  const pageContext = getPageContext();
-  const page = ctx.renderPage(Component => props => (
-    <Component pageContext={pageContext} {...props} />
-  ));
+  // Render app and page and get the context of the page with collected side effects.
+  let pageContext;
+  const page = ctx.renderPage(Component => {
+    const WrappedComponent = props => {
+      pageContext = props.pageContext;
+      return <Component {...props} />;
+    };
+
+    WrappedComponent.propTypes = {
+      pageContext: PropTypes.object.isRequired,
+    };
+
+    return WrappedComponent;
+  });
 
   let css = pageContext.sheetsRegistry.toString();
   if (process.env.NODE_ENV === 'production') {
